chore: update to 1.19.4

+++ src/main/java/com/github/klikli_dev/occultism/common/entity/familiar/FamiliarEntity.java
@@ -301,8 +301,8 @@ public abstract class FamiliarEntity extends PathfinderMob implements IFamiliar 
 
         private BlockPos randomNearby(BlockPos pos) {
             var rand = this.entity.getRandom();
-            return pos.offset(Mth.nextDouble(rand, -3, 3), Mth.nextDouble(rand, -1, 1),
-                    Mth.nextDouble(rand, -3, 3));
+            return pos.offset(Mth.nextInt(rand, -3, 3), Mth.nextInt(rand, -1, 1),
+                    Mth.nextInt(rand, -3, 3));
         }
     }
